ci and requested changes
