+++ sdk/src/common/BUILD
@@ -58,17 +58,6 @@ cc_library(
 )
 
 cc_library(
-    name = "container",
-    srcs = [
-        "container.cc",
-    ],
-    deps = [
-        "//api",
-        "//sdk:headers",
-    ],
-)
-
-cc_library(
     name = "global_log_handler",
     srcs = [
         "global_log_handler.cc",
