+++ sdk/src/common/CMakeLists.txt
@@ -1,8 +1,8 @@
 # Copyright The OpenTelemetry Authors
 # SPDX-License-Identifier: Apache-2.0
 
-set(COMMON_SRCS random.cc global_log_handler.cc env_variables.cc container.cc
-                base64.cc disabled.cc)
+set(COMMON_SRCS random.cc global_log_handler.cc env_variables.cc base64.cc
+                disabled.cc)
 if(WIN32)
   list(APPEND COMMON_SRCS platform/fork_windows.cc)
 else()
