+++ sdk/src/resource/BUILD
@@ -10,7 +10,6 @@ cc_library(
     deps = [
         "//api",
         "//sdk:headers",
-        "//sdk/src/common:container",
         "//sdk/src/common:env_variables",
     ],
 )
