+++ sdk/src/resource/CMakeLists.txt
@@ -1,8 +1,8 @@
 # Copyright The OpenTelemetry Authors
 # SPDX-License-Identifier: Apache-2.0
 
-add_library(opentelemetry_resources resource.cc resource_detector.cc 
-                                    container_detector.cc)
+add_library(opentelemetry_resources resource.cc resource_detector.cc
+                                    container.cc container_detector.cc)
 
 set_target_properties(opentelemetry_resources PROPERTIES EXPORT_NAME resources)
 set_target_version(opentelemetry_resources)
